M2-5059: Create separated file for monthly tests
