+++ src/entities/event/model/operations/ScheduledDateCalculator.test.js
@@ -1,4 +1,4 @@
-import { addDays, addMonths, startOfDay, subDays, subMonths } from 'date-fns';
+import { addDays, startOfDay, subDays } from 'date-fns';
 import ScheduledDateCalculator from './ScheduledDateCalculator';
 
 const now = new Date(2024, 0, 25);
@@ -428,190 +428,6 @@ describe('ScheduledDateCalculator', () => {
       const event = getWeeklyEvent();
       event.availability.startDate = addDays(startOfDay(now), 1);
       event.availability.endDate = addDays(startOfDay(now), 5);
-      event.selectedDate = subDays(startOfDay(now), 1);
-
-      const resultDate = ScheduledDateCalculator.calculate(event, false);
-
-      expect(resultDate).toEqual(null);
-    });
-  });
-
-  describe('Test Monthly events', () => {
-    const getMonthlyEvent = () => {
-      const eventAvailability = {
-        availabilityType: 'ScheduledAccess',
-        periodicityType: 'MONTHLY',
-        timeFrom: { hours: 14, minutes: 0 },
-        timeTo: { hours: 18, minutes: 30 },
-        startDate: null,
-        endDate: null,
-      };
-
-      const scheduleEvent = {
-        id: 'eventTestId',
-        entityId: 'entityTestId',
-        availability: eventAvailability,
-        selectedDate: null,
-        scheduledAt: null,
-      };
-
-      return scheduleEvent;
-    };
-
-    it("Should return today's start of day when selected day is today and is covered by start-end dates and timeFrom is reset", () => {
-      const event = getMonthlyEvent();
-      event.availability.startDate = subDays(startOfDay(now), 2);
-      event.availability.endDate = addDays(startOfDay(now), 2);
-      event.availability.timeFrom = { hours: 0, minutes: 0 };
-      event.selectedDate = startOfDay(now);
-
-      const resultDate = ScheduledDateCalculator.calculate(event, false);
-
-      const expectedDate = startOfDay(now);
-
-      expect(resultDate).toEqual(expectedDate);
-    });
-
-    it('Should return today with timeFrom when selected day is today and covered by start-end dates and timeFrom is set', () => {
-      const event = getMonthlyEvent();
-      event.availability.startDate = subDays(startOfDay(now), 2);
-      event.availability.endDate = addDays(startOfDay(now), 2);
-      event.selectedDate = startOfDay(now);
-
-      const resultDate = ScheduledDateCalculator.calculate(event, false);
-
-      const expectedDate = startOfDay(now);
-      expectedDate.setHours(event.availability.timeFrom.hours);
-      expectedDate.setMinutes(event.availability.timeFrom.minutes);
-
-      expect(resultDate).toEqual(expectedDate);
-    });
-
-    it('Should return date today with timeFrom when selected day is (today + or - 1 month) and timeFrom is set and date start/end are (today -/+ 2 months)', () => {
-      const event = getMonthlyEvent();
-      event.availability.startDate = subMonths(startOfDay(now), 2);
-      event.availability.endDate = addMonths(startOfDay(now), 2);
-      event.selectedDate = addMonths(startOfDay(now), 1);
-
-      let resultDate = ScheduledDateCalculator.calculate(event, false);
-
-      const expectedDate = startOfDay(now);
-      expectedDate.setHours(event.availability.timeFrom.hours);
-      expectedDate.setMinutes(event.availability.timeFrom.minutes);
-
-      expect(resultDate).toEqual(expectedDate);
-
-      event.selectedDate = subMonths(startOfDay(now), 1);
-
-      resultDate = ScheduledDateCalculator.calculate(event, false);
-
-      expect(resultDate).toEqual(expectedDate);
-    });
-
-    it('Should return a date equal to (tomorrow - 1 month) with set/not set timeFrom when selected date is tomorrow and start/end dates are (today +1/+25 days)', () => {
-      const event = getMonthlyEvent();
-      event.availability.startDate = addDays(startOfDay(now), 1);
-      event.availability.endDate = addDays(startOfDay(now), 25);
-      event.selectedDate = addDays(startOfDay(now), 1);
-
-      let resultDate = ScheduledDateCalculator.calculate(event, false);
-
-      let expectedDate = subMonths(addDays(startOfDay(now), 1), 1);
-      expectedDate.setHours(event.availability.timeFrom.hours);
-      expectedDate.setMinutes(event.availability.timeFrom.minutes);
-
-      expect(resultDate).toEqual(expectedDate);
-
-      event.availability.timeFrom = { hours: 0, minutes: 0 };
-
-      resultDate = ScheduledDateCalculator.calculate(event, false);
-
-      expectedDate = subMonths(addDays(startOfDay(now), 1), 1);
-
-      expect(resultDate).toEqual(expectedDate);
-    });
-
-    it('Should return a date equal to endDate with set/not set timeFrom when selected date is yesterday and start/end dates are (today -25/-1 days)', () => {
-      const event = getMonthlyEvent();
-      event.availability.startDate = subDays(startOfDay(now), 25);
-      event.availability.endDate = subDays(startOfDay(now), 1);
-      event.selectedDate = subDays(startOfDay(now), 1);
-
-      let resultDate = ScheduledDateCalculator.calculate(event, false);
-
-      let expectedDate = subDays(startOfDay(now), 1);
-      expectedDate.setHours(event.availability.timeFrom.hours);
-      expectedDate.setMinutes(event.availability.timeFrom.minutes);
-
-      expect(resultDate).toEqual(expectedDate);
-
-      event.availability.timeFrom = { hours: 0, minutes: 0 };
-
-      resultDate = ScheduledDateCalculator.calculate(event, false);
-
-      expectedDate = subDays(startOfDay(now), 1);
-
-      expect(resultDate).toEqual(expectedDate);
-    });
-
-    it('Should return a date equal to endDate with set/not set timeFrom when selected date is yesterday and start/end dates are (today -55/-1 days)', () => {
-      // -2 todo
-      const event = getMonthlyEvent();
-      event.availability.startDate = subDays(startOfDay(now), 55);
-      event.availability.endDate = subDays(startOfDay(now), 1);
-      event.selectedDate = subDays(startOfDay(now), 1);
-
-      let resultDate = ScheduledDateCalculator.calculate(event, false);
-
-      let expectedDate = subDays(startOfDay(now), 1);
-      expectedDate.setHours(event.availability.timeFrom.hours);
-      expectedDate.setMinutes(event.availability.timeFrom.minutes);
-
-      expect(resultDate).toEqual(expectedDate);
-
-      event.availability.timeFrom = { hours: 0, minutes: 0 };
-
-      resultDate = ScheduledDateCalculator.calculate(event, false);
-
-      expectedDate = subDays(startOfDay(now), 1);
-
-      expect(resultDate).toEqual(expectedDate);
-    });
-
-    it('Should return (today + 10 days - 1 month days with timeFrom) when timeFrom is set and selected date is (today + 10 days) and start/end dates -/+ 50 days', () => {
-      const event = getMonthlyEvent();
-      event.availability.startDate = subDays(startOfDay(now), 50);
-      event.availability.endDate = addDays(startOfDay(now), 50);
-      event.selectedDate = addDays(startOfDay(now), 10);
-
-      const resultDate = ScheduledDateCalculator.calculate(event, false);
-
-      const expectedDate = subMonths(addDays(startOfDay(now), 10), 1);
-      expectedDate.setHours(event.availability.timeFrom.hours);
-      expectedDate.setMinutes(event.availability.timeFrom.minutes);
-
-      expect(resultDate).toEqual(expectedDate);
-    });
-
-    it('Should return (today - 4 days with timeFrom) when timeFrom is set and selected date is (today - 4 days) and start/end dates -/+ 50 days', () => {
-      const event = getMonthlyEvent();
-      event.availability.startDate = subDays(startOfDay(now), 50);
-      event.availability.endDate = addDays(startOfDay(now), 50);
-      event.selectedDate = subDays(startOfDay(now), 4);
-
-      const resultDate = ScheduledDateCalculator.calculate(event, false);
-
-      const expectedDate = subDays(startOfDay(now), 4);
-      expectedDate.setHours(event.availability.timeFrom.hours);
-      expectedDate.setMinutes(event.availability.timeFrom.minutes);
-
-      expect(resultDate).toEqual(expectedDate);
-    });
-
-    it('Should return null when selected date is (today - 1 days) and start/end date earlier than (today -1 day)', () => {
-      const event = getMonthlyEvent();
-      event.availability.startDate = subDays(startOfDay(now), 56);
-      event.availability.endDate = subDays(startOfDay(now), 3);
       event.selectedDate = subDays(startOfDay(now), 1);
 
       const resultDate = ScheduledDateCalculator.calculate(event, false);
